2mm panel stls, modified magnet receiver, modified CAD
Created stls for 2mm panels that came with Fysetc V0.1 kit and modified magnet receiver to work with hbo's pivot bed mounts (front bed clamp would collide otherwise). The Zero Panels CAD file was modified so that the hole for the M2 lock screws for the top and bottom hinges are linked to the length of the clipped region. Additionally, the clip region for top fixed hinge was offset to clear the front idlers on stock v0.2.